19/04/2020 02:20 fix bug taux d'absence (double and , apres la vergule)

+++ Assiduite/Pages/Etudiants/Index.cshtml.cs
@@ -119,7 +119,7 @@ namespace Assiduite.Pages.Etudiants
 
                     TotalAbs = await _db.presence.Where(e => e.Id_Etudiant_Pres == _student.Id_Etudiant && e.Etat_Pres == 1).CountAsync();
 
-                    double PourcentageAbs = (TotalAbs / TotalSeance) * 100;
+                    double PourcentageAbs = ((double)TotalAbs / (double)TotalSeance) * 100;
                     
                     _studentAbs.Add(new StudentAbs(_student, TotalAbs, PourcentageAbs));
                 }
